update bad_host test

+++ core/src/util/test_connection.rs
@@ -133,7 +133,7 @@ mod test {
         };
         assert!(!result);
         assert!(unsafe {
-            input_text_to_string_w(buffer.as_mut_ptr(), buffer_len as usize)
+            input_text_to_string_w(buffer.as_mut_ptr(), buffer_len)
                 .to_lowercase()
                 .contains("server selection timeout")
         });
